requirements.txt
pytest==9.0.2
requests==2.32.5
jsonschema==4.25.1
pytest-html==4.1.1
allure-pytest==2.15.0
pytest-xdist==3.8.0
python-dotenv==1.1.1